requirements.txt
async-timeout==4.0.3
loguru==0.7.2
py-dotenv==0.1
python-dotenv==1.0.1
python-socks==2.4.4
websockets==12.0
websockets_proxy==0.1.2
